+++ examples/grid.py
@@ -1,13 +1,12 @@
 from textual.app import App
-from textual import events
 from textual.widgets import Placeholder
 
 
 class GridTest(App):
-    async def on_mount(self, event: events.Mount) -> None:
+    async def on_mount(self) -> None:
         """Make a simple grid arrangement."""
 
-        grid = await self.view.dock_grid(edge="left", size=70, name="left")
+        grid = await self.view.dock_grid(edge="left", name="left")
 
         grid.add_column(fraction=1, name="left", min_size=20)
         grid.add_column(size=30, name="center")