+++ src/textual/app.py
@@ -9,8 +9,8 @@ import warnings
 from rich.control import Control
 import rich.repr
 from rich.screen import Screen
-from rich import get_console
 from rich.console import Console, RenderableType
+from rich.measure import Measurement
 from rich.traceback import Traceback
 
 from . import events
@@ -19,14 +19,14 @@ from ._animator import Animator
 from .binding import Bindings, NoBinding
 from .geometry import Offset, Region
 from . import log
+from ._callback import invoke
 from ._context import active_app
 from ._event_broker import extract_handler_actions, NoHandler
-from ._types import MessageTarget
 from .driver import Driver
 from .layouts.dock import DockLayout, Dock
 from ._linux_driver import LinuxDriver
 from .message_pump import MessagePump
-from .message import Message
+from ._profile import timer
 from .view import View
 from .views import DockView
 from .widget import Widget, Widget, Reactive
@@ -48,17 +48,7 @@ else:
     uvloop.install()
 
 
-class PanicMessage(Message):
-    def __init__(self, sender: MessageTarget, traceback: Traceback) -> None:
-        self.traceback = traceback
-        super().__init__(sender)
-
-
 class ActionError(Exception):
-    pass
-
-
-class ShutdownError(Exception):
     pass
 
 
@@ -111,7 +101,7 @@ class App(MessagePump):
         self.log_file = open(log, "wt") if log else None
         self.log_verbosity = log_verbosity
 
-        self.bindings.bind("ctrl+c", "quit", show=False)
+        self.bindings.bind("ctrl+c", "quit", show=False, allow_forward=False)
         self._refresh_required = False
 
         super().__init__()
@@ -135,6 +125,12 @@ class App(MessagePump):
         return self._view_stack[-1]
 
     def log(self, *args: Any, verbosity: int = 1) -> None:
+        """Write to logs.
+
+        Args:
+            *args (Any): Positional arguments are converted to string and written to logs.
+            verbosity (int, optional): Verbosity level 0-3. Defaults to 1.
+        """
         try:
             if self.log_file and verbosity <= self.log_verbosity:
                 output = f" ".join(str(arg) for arg in args)
@@ -151,6 +147,15 @@ class App(MessagePump):
         show: bool = True,
         key_display: str | None = None,
     ) -> None:
+        """Bind a key to an action.
+
+        Args:
+            keys (str): A comma separated list of keys, i.e.
+            action (str): Action to bind to.
+            description (str, optional): Short description of action. Defaults to "".
+            show (bool, optional): Show key in UI. Defaults to True.
+            key_display (str, optional): Replacement text for key, or None to use default. Defaults to None.
+        """
         self.bindings.bind(
             keys, action, description, show=show, key_display=key_display
         )
@@ -180,17 +185,17 @@ class App(MessagePump):
     async def push_view(self, view: ViewType) -> ViewType:
         self.register(view, self)
         self._view_stack.append(view)
-        # await view.post_message(events.Mount(sender=self))
         return view
 
-    def on_keyboard_interupt(self) -> None:
-        loop = asyncio.get_event_loop()
-        event = events.ShutdownRequest(sender=self)
-        asyncio.run_coroutine_threadsafe(self.post_message(event), loop=loop)
-
     async def set_focus(self, widget: Widget | None) -> None:
+        """Focus (or unfocus) a widget. A focused widget will receive key events first.
+
+        Args:
+            widget (Widget): [description]
+        """
         log("set_focus", widget)
         if widget == self.focused:
+            # Widget is already focused
             return
 
         if widget is None:
@@ -223,7 +228,11 @@ class App(MessagePump):
                     self.mouse_over = widget
 
     async def capture_mouse(self, widget: Widget | None) -> None:
-        """Send all Mouse events to a given widget."""
+        """Send all mouse events to the given widget, disable mouse capture.
+
+        Args:
+            widget (Widget | None): If a widget, capture mouse event, or None to end mouse capture.
+        """
         if widget == self.mouse_captured:
             return
         if self.mouse_captured is not None:
@@ -251,22 +260,24 @@ class App(MessagePump):
 
     async def process_messages(self) -> None:
         active_app.set(self)
-        driver = self._driver = self.driver_class(self.console, self)
 
         log("---")
         log(f"driver={self.driver_class}")
 
-        await self.dispatch_message(events.Load(sender=self))
+        load_event = events.Load(sender=self)
+        await self.dispatch_message(load_event)
         await self.post_message(events.Mount(self))
         await self.push_view(DockView())
 
+        # Wait for the load event to be processed, so we don't go in to application mode beforehand
+        await load_event.wait()
+
+        driver = self._driver = self.driver_class(self.console, self)
         try:
             driver.start_application_mode()
         except Exception:
             self.console.print_exception()
         else:
-            traceback: Traceback | None = None
-
             try:
                 self.title = self._title
                 self.refresh()
@@ -274,24 +285,15 @@ class App(MessagePump):
                 await super().process_messages()
                 log("PROCESS END")
                 await self.animator.stop()
-
                 await self.close_all()
 
-                # while self.children:
-                #     child = self.children.pop()
-                #     log(f"closing {child}")
-                #     await child.close_messages()
-
-                # while self._view_stack:
-                #     view = self._view_stack.pop()
-                #     await view.close_messages()
             except Exception:
                 self.panic()
             finally:
                 driver.stop_application_mode()
                 if self._exit_renderables:
-                    for traceback in self._exit_renderables:
-                        self.error_console.print(traceback)
+                    for renderable in self._exit_renderables:
+                        self.error_console.print(renderable)
                 if self.log_file is not None:
                     self.log_file.close()
 
@@ -346,10 +348,42 @@ class App(MessagePump):
             except Exception:
                 self.panic()
 
+    def measure(self, renderable: RenderableType, max_width=100_000) -> int:
+        """Get the optimal width for a widget or renderable.
+
+        Args:
+            renderable (RenderableType): A renderable (including Widget)
+            max_width ([type], optional): Maximum width. Defaults to 100_000.
+
+        Returns:
+            int: Number of cells required to render.
+        """
+        measurement = Measurement.get(
+            self.console, self.console.options.update(max_width=max_width), renderable
+        )
+        return measurement.maximum
+
     def get_widget_at(self, x: int, y: int) -> tuple[Widget, Region]:
+        """Get the widget under the given coordinates.
+
+        Args:
+            x (int): X Coord.
+            y (int): Y Coord.
+
+        Returns:
+            tuple[Widget, Region]: The widget and the widget's screen region.
+        """
         return self.view.get_widget_at(x, y)
 
     async def press(self, key: str) -> bool:
+        """Handle a key press.
+
+        Args:
+            key (str): A key
+
+        Returns:
+            bool: True if the key was handled by a binding, otherwise False
+        """
         try:
             binding = self.bindings.get_key(key)
         except NoBinding:
@@ -359,23 +393,28 @@ class App(MessagePump):
         return True
 
     async def on_event(self, event: events.Event) -> None:
-        if isinstance(event, events.Key):
-            if await self.press(event.key):
-                return
-            await super().on_event(event)
-
-        if isinstance(event, events.InputEvent):
+        # Handle input events that haven't been forwarded
+        # If the event has been forwaded it may have bubbled up back to the App
+        if isinstance(event, events.InputEvent) and not event.is_forwarded:
             if isinstance(event, events.MouseEvent):
+                # Record current mouse position on App
                 self.mouse_position = Offset(event.x, event.y)
             if isinstance(event, events.Key) and self.focused is not None:
-                await self.focused.forward_event(event)
-            await self.view.forward_event(event)
+                # Key events are sent direct to focused widget
+                if self.bindings.allow_forward(event.key):
+                    await self.focused.forward_event(event)
+                else:
+                    # Key has allow_forward=False which disallows forward to focused widget
+                    await super().on_event(event)
+            else:
+                # Forward the event to the view
+                await self.view.forward_event(event)
         else:
             await super().on_event(event)
 
-    async def on_idle(self, event: events.Idle) -> None:
-        if self.view.check_layout():
-            await self.view.refresh_layout()
+    # async def on_idle(self, event: events.Idle) -> None:
+    #     if self.view.check_layout():
+    #         await self.view.refresh_layout()
 
     async def action(
         self,
@@ -407,8 +446,8 @@ class App(MessagePump):
         _rich_traceback_guard = True
         method_name = f"action_{action_name}"
         method = getattr(namespace, method_name, None)
-        if method is not None:
-            await method(*params)
+        if callable(method):
+            await invoke(method, *params)
 
     async def broker_event(
         self, event_name: str, event: events.Event, default_namespace: object | None
@@ -430,6 +469,10 @@ class App(MessagePump):
         else:
             return False
         return True
+
+    async def on_key(self, event: events.Key) -> None:
+        # self.log("App.on_key")
+        await self.press(event.key)
 
     async def on_shutdown_request(self, event: events.ShutdownRequest) -> None:
         log("shutdown request")